# swagger.json
{"openapi":"3.0.0","info":{"title":"SMARTHOME-CHBAN API Document","description":"The Smart Home Backend API Description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"SignUpRequestDto":{"type":"object","properties":{"fname":{"type":"string","example":"Firstname"},"lname":{"type":"string","example":"Lastname"},"username":{"type":"string","example":"Username"},"password":{"type":"string","example":"password"}},"required":["fname","lname","username","password"]},"User":{"type":"object","properties":{"fname":{"type":"string","example":"Firstname"},"lname":{"type":"string","example":"Lastname"},"createTime":{"format":"date-time","type":"string","example":"2021-05-08 16:03:38"},"firstAccess":{"format":"date-time","type":"string","example":"2021-05-08 16:03:38"},"lastAccess":{"format":"date-time","type":"string","example":"2021-05-08 16:03:38"},"username":{"type":"string","example":"USERNAME"},"password":{"type":"string","example":"PASSWORD"},"raspberryPis":{"example":[{}],"type":"array","items":{"type":"string"}}},"required":["fname","lname","createTime","firstAccess","lastAccess","username","password","raspberryPis"]},"SignInRequestDto":{"type":"object","properties":{"username":{"type":"string","example":"username"},"password":{"type":"string","example":"password"}},"required":["username","password"]},"SignInResponseDto":{"type":"object","properties":{"id":{"type":"string","example":"1"},"fname":{"type":"string","example":"Firstname"},"lname":{"type":"string","example":"Lastname"},"token":{"type":"string","example":"asdasdasdasdsdlhulfkwepufbdkmckanfwk.sdfhekgfuehflieyguehifjdskluwbusdpijaknjniofuwe.sdfglihvsnchaskbucbu"}},"required":["id","fname","lname","token"]},"GetAllUserResponseDto":{"type":"object","properties":{"id":{"type":"string","example":"1"},"fname":{"type":"string","example":"Firstname"},"lname":{"type":"string","example":"Lastname"},"piList":{"example":[{}],"type":"array","items":{"type":"string"}}},"required":["id","fname","lname","piList"]},"AddRaspberryPiRequestDto":{"type":"object","properties":{"userId":{"type":"string","example":"1"},"piName":{"type":"string","example":"Raspberry Pi Name"},"status":{"type":"number","example":1},"positionX":{"type":"string","example":"120.0054154"},"positionY":{"type":"string","example":"12.000"}},"required":["userId","piName","status","positionX","positionY"]},"RaspberryPi":{"type":"object","properties":{"piId":{"type":"string","example":"1"},"name":{"type":"string","example":"PI_name"},"createTime":{"format":"date-time","type":"string","example":"2021-05-08 16:03:38"},"status":{"type":"number","example":1},"updateTime":{"format":"date-time","type":"string","example":"2021-05-08 16:03:38"},"userId":{"type":"string","example":"2"},"positionX":{"type":"string","example":"2"},"positionY":{"type":"string","example":"2"},"otp":{"type":"string","example":"111000"},"deviceLights":{"example":[{}],"type":"array","items":{"type":"string"}},"deviceDhts":{"example":[{}],"type":"array","items":{"type":"string"}}},"required":["piId","name","createTime","status","updateTime","userId","positionX","positionY","otp","deviceLights","deviceDhts"]},"InstallRaspberryPiRequestDto":{"type":"object","properties":{"piId":{"type":"string","example":"1"},"name":{"type":"string","example":"หน้าบ้าน"},"otp":{"type":"string","example":"123456"}},"required":["piId","name","otp"]},"ConfirmOTPRaspberryPiRequestDto":{"type":"object","properties":{"piId":{"type":"string","example":"1"},"otp":{"type":"string","example":"123456"}},"required":["piId","otp"]},"GetAllRaspberryPiByUserId":{"type":"object","properties":{"piId":{"type":"string","example":"1"},"name":{"type":"string","example":"Raspberry Pi Name"},"status":{"type":"number","example":1},"dhtList":{"example":[{}],"type":"array","items":{"type":"string"}},"lightList":{"example":[{}],"type":"array","items":{"type":"string"}},"positionX":{"type":"string","example":"120.0054154"},"positionY":{"type":"string","example":"12.000"}},"required":["piId","name","status","dhtList","lightList","positionX","positionY"]},"UpdateRaspberryPiStatusRequestDto":{"type":"object","properties":{"piId":{"type":"string","example":"1"},"status":{"type":"number","example":1}},"required":["piId","status"]},"AddDHTRequestDto":{"type":"object","properties":{"piId":{"type":"string","example":"1"},"name":{"type":"string","example":"DHT Name"},"isOnline":{"type":"number","example":1},"positionX":{"type":"string","example":"120.0054154"},"positionY":{"type":"string","example":"12.000"}},"required":["piId","name","isOnline","positionX","positionY"]},"DeviceDht":{"type":"object","properties":{"dhtId":{"type":"string","example":"1"},"name":{"type":"string","example":"DHT Name"},"status":{"type":"number","example":1},"isOnline":{"type":"number","example":1},"temperature":{"type":"string","example":"36.0"},"moisture":{"type":"string","example":"20.0"},"swLastUpdate":{"format":"date-time","type":"string","example":"2021-05-08 16:03:38"},"hwLastUpdate":{"format":"date-time","type":"string","example":"2021-05-08 16:03:38"},"piId":{"type":"string","example":"36"},"positionX":{"type":"string","example":"1.000"},"positionY":{"type":"string","example":"2.00"},"otp":{"type":"string","example":"000000"}},"required":["dhtId","name","status","isOnline","temperature","moisture","swLastUpdate","hwLastUpdate","piId","positionX","positionY","otp"]},"InstallDHTRequestDto":{"type":"object","properties":{"dhtId":{"type":"string","example":"1"},"name":{"type":"string","example":"Test DHT"},"otp":{"type":"string","example":"123456"}},"required":["dhtId","name","otp"]},"ConfirmOTPDhtRequestDto":{"type":"object","properties":{"dhtId":{"type":"string","example":"1"},"otp":{"type":"string","example":"123456"}},"required":["dhtId","otp"]},"DHTGetValueResponseDto":{"type":"object","properties":{"dhtId":{"type":"string","example":"1"},"piId":{"type":"string","example":"1"},"name":{"type":"string","example":"DHT name"},"status":{"type":"number","example":1},"isOnline":{"type":"number","example":1},"temperature":{"type":"string","example":"100"},"moisture":{"type":"string","example":"100"},"positionX":{"type":"string","example":"2120.0"},"positionY":{"type":"string","example":"123"}},"required":["dhtId","piId","name","status","isOnline","temperature","moisture","positionX","positionY"]},"UpdateDHTStatusRequestDto":{"type":"object","properties":{"id":{"type":"string","example":"1"},"status":{"type":"number","example":1},"isOnline":{"type":"number","example":1}},"required":["id","status","isOnline"]},"UpdateDHTValueRequestDto":{"type":"object","properties":{"id":{"type":"string","example":"1"},"temperature":{"type":"string","example":"36.9"},"moisture":{"type":"string","example":"110"}},"required":["id","temperature","moisture"]}}},"paths":{"/auth/sign-up":{"post":{"operationId":"AuthController_signUp","summary":"Sign up","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpRequestDto"}}}},"responses":{"200":{"description":"Sign up success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["Auth"]}},"/auth/sign-in":{"post":{"operationId":"AuthController_signIn","summary":"Sign in","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInRequestDto"}}}},"responses":{"200":{"description":"Sign in success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInResponseDto"}}}}},"tags":["Auth"]}},"/auth/sign-out":{"post":{"operationId":"AuthController_signOut","summary":"Sign out","parameters":[],"responses":{"200":{"description":"Sign out success","content":{"application/json":{"schema":{"type":"boolean"}}}}},"tags":["Auth"]}},"/user/get-all-user":{"get":{"operationId":"UserController_getAllUser","summary":"Get all users list","parameters":[],"responses":{"200":{"description":"Get all users success list","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetAllUserResponseDto"}}}}}},"tags":["User"]}},"/user/get-user/{id}":{"get":{"operationId":"UserController_getAllUserById","summary":"Get user by user ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Get user by user ID success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetAllUserResponseDto"}}}}},"tags":["User"]}},"/device/pi/add-pi":{"post":{"operationId":"DeviceController_addRaspberryPi","summary":"Add Raspberry Pi","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddRaspberryPiRequestDto"}}}},"responses":{"200":{"description":"Add Raspberry Pi success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RaspberryPi"}}}}},"tags":["Device"]}},"/device/pi/install":{"post":{"operationId":"DeviceController_installRaspberryPi","summary":"Install Raspberry Pi","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InstallRaspberryPiRequestDto"}}}},"responses":{"200":{"description":"Install Raspberry Pi success","content":{"application/json":{"schema":{"type":"boolean"}}}}},"tags":["Device"]}},"/device/pi/confirm-otp":{"post":{"operationId":"DeviceController_confirmOtpRaspberryPi","summary":"Confirm OTP Raspberry Pi","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmOTPRaspberryPiRequestDto"}}}},"responses":{"200":{"description":"Confirm OTP Raspberry Pi success","content":{"application/json":{"schema":{"type":"boolean"}}}}},"tags":["Device"]}},"/device/pi/get-pi-by-user-id/{userId}":{"get":{"operationId":"DeviceController_getAllRaspberryPiByUserId","summary":"Get Raspberry Pi by user ID","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Get Raspberry Pi by user ID success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetAllRaspberryPiByUserId"}}}}},"tags":["Device"]}},"/device/pi/get-pi-by-id/{piId}":{"get":{"operationId":"DeviceController_getRaspberryPiById","summary":"Get Raspberry Pi by ID","parameters":[{"name":"piId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Get Raspberry Pi by ID success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetAllRaspberryPiByUserId"}}}}},"tags":["Device"]}},"/device/pi/update-status":{"put":{"operationId":"DeviceController_updatePiStatus","summary":"Update Raspberry Pi status","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRaspberryPiStatusRequestDto"}}}},"responses":{"200":{"description":"Update Raspberry Pi status success","content":{"application/json":{"schema":{"type":"boolean"}}}}},"tags":["Device"]}},"/device/dht/add-dht":{"post":{"operationId":"DeviceController_addDHT","summary":"Add DHT","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddDHTRequestDto"}}}},"responses":{"200":{"description":"Add DHT success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeviceDht"}}}}},"tags":["Device"]}},"/device/dht/install":{"post":{"operationId":"DeviceController_installDHT","summary":"Install DHT","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InstallDHTRequestDto"}}}},"responses":{"200":{"description":"Install DHT success","content":{"application/json":{"schema":{"type":"boolean"}}}}},"tags":["Device"]}},"/device/dht/confirm-otp":{"post":{"operationId":"DeviceController_confirmOtpDHT","summary":"Confirm OTP DHT","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmOTPDhtRequestDto"}}}},"responses":{"200":{"description":"Confirm OTP DHT success","content":{"application/json":{"schema":{"type":"boolean"}}}}},"tags":["Device"]}},"/device/dht/get-all-by-pi-id/{piId}":{"get":{"operationId":"DeviceController_getDhtByPiId","summary":"Get all DHT by Raspberry Pi ID","parameters":[{"name":"piId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Get all DHT by Raspberry Pi ID success","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DHTGetValueResponseDto"}}}}}},"tags":["Device"]}},"/device/dht/get-all-by-user-id/{userId}":{"get":{"operationId":"DeviceController_getDhtByUserId","summary":"Get all DHT by user ID","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Get all DHT by user ID success","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DHTGetValueResponseDto"}}}}}},"tags":["Device"]}},"/device/dht/update-status":{"put":{"operationId":"DeviceController_updateDHTStatus","summary":"Update DHT status","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateDHTStatusRequestDto"}}}},"responses":{"200":{"description":"Update DHT status success","content":{"application/json":{"schema":{"type":"boolean"}}}}},"tags":["Device"]}},"/device/dht/update-value":{"put":{"operationId":"DeviceController_updateDHTValue","summary":"Update DHT value","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateDHTValueRequestDto"}}}},"responses":{"200":{"description":"Update DHT value success","content":{"application/json":{"schema":{"type":"boolean"}}}}},"tags":["Device"]}}}}